Integrate project source code
