--- QLHS/QLHocSinh/admin/index.html
@@ -1,25 +1,0 @@
-22 lines (19 sloc)  823 Bytes
-
-
-{% extends 'admin/master.html' %}
-
-{% block body %}
-
-{% if current_user.is_authenticated %}
-<h1 class="text-success">Chào mừng bạn đã quay lại, {{ current_user.name }}</h1>
-{% else %}
-<h1 class="text-center text-info">ĐĂNG NHẬP QUẢN TRỊ</h1>
-
-<form method="post" action="/login?next=/admin">
-  <div class="mb-3 mt-3">
-    <label for="username" class="form-label">Tên đăng nhập:</label>
-    <input type="text" class="form-control" id="username" placeholder="Tên đăng nhập" name="username">
-  </div>
-  <div class="mb-3">
-    <label for="pwd" class="form-label">Mật khẩu:</label>
-    <input type="password" class="form-control" id="pwd" placeholder="Mật khẩu" name="password">
-  </div>
-  <button type="submit" class="btn btn-primary">Đăng nhập</button>
-</form>
-{% endif %}
-{% endblock %}
