--- QLHS/QLHocSinh/admin/stats.html
@@ -1,6 +1,0 @@
-<!--{% extends 'admin/master.html' %}-->
-
-<!--{% block body %}-->
-<h1 class="text-center text-danger">THỐNG KÊ BÁO CÁO</h1>
-
-<!--{% endblock %}-->
