--- QLHS/QLHocSinh/templates/arrange-class.html
@@ -1,6 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-<h1 class="text-center heading">ĐIỀU CHỈNH LỚP HỌC</h1>
-
-{% endblock %}
