--- QLHS/QLHocSinh/templates/base.html
@@ -1,71 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Trang chủ</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
-          integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"
-            integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
-            crossorigin="anonymous"></script>
-</head>
-<body class="bg-info">
-<h1 align="center"></h1>
-<nav class="navbar navbar-expand-sm bg-dark navbar-dark">
-    <div class="container-fluid">
-        <a class="navbar-brand" href="#">TRƯỜNG TPHT ABC</a>
-        <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#collapsibleNavbar">
-            <span class="navbar-toggler-icon"></span>
-        </button>
-        <div class="collapse navbar-collapse" id="collapsibleNavbar">
-            <ul class="navbar-nav me-auto">
-                <li class="nav-item">
-                    <a class="nav-link" href="/">Trang chủ</a>
-                </li>
-                {% if current_user.is_authenticated %}
-                <li class="nav-item">
-                    <a class="nav-link text-danger" href="/">
-                        <img src="{{ current_user.avatar }}" width="40" class="rounded-circle" />{{ current_user.name }}
-                    </a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="/add-students">Tiếp nhận học sinh</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="/setup-class">Lập danh sách lớp</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="/arrange-class">Điều chỉnh lớp học</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="/students-marks">Quản lý điểm</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="/logout">Đăng xuất</a>
-                </li>
-                {% else %}
-                <li class="nav-item">
-                        <a class="nav-link" href="/login">Đăng nhập</a>
-                </li>
-                {% endif %}
-            </ul>
-             <form class="d-flex" action="/">
-                <input class="form-control me-2" name="kw" type="text" placeholder="Nhập tên...">
-                <button class="btn btn-primary" type="submit">Tìm</button>
-             </form>
-        </div>
-    </div>
-</nav>
-
-<div class="container">
-{% block content %}
-{% endblock %}
-</div>
-
-<div class="mt-4 p-5 bg-primary text-white rounded">
-  <h1>E-commerce Website</h1>
-  <p>Trường THPT ABC &copy; 2023</p>
-</div>
-
-</body>
-</html>
