--- QLHS/QLHocSinh/templates/index.html
@@ -1,35 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-<!-- Carousel -->
-<div id="demo" class="carousel slide" data-bs-ride="carousel">
-
-    <!-- Indicators/dots -->
-    <div class="carousel-indicators">
-        <button type="button" data-bs-target="#demo" data-bs-slide-to="0" class="active"></button>
-        <button type="button" data-bs-target="#demo" data-bs-slide-to="1"></button>
-        <button type="button" data-bs-target="#demo" data-bs-slide-to="2"></button>
-    </div>
-
-    <!-- The slideshow/carousel -->
-    <div class="carousel-inner">
-        <div class="carousel-item active">
-            <img src="https://res.cloudinary.com/dkcq9dgui/image/upload/v1678842721/cld-sample-3.jpg" class="d-block w-100">
-        </div>
-        <div class="carousel-item">
-            <img src="https://res.cloudinary.com/dkcq9dgui/image/upload/v1678842721/cld-sample-4.jpg" class="d-block w-100">
-        </div>
-        <div class="carousel-item">
-            <img src="https://res.cloudinary.com/dkcq9dgui/image/upload/v1678842706/samples/ecommerce/leather-bag-gray.jpg" class="d-block w-100">
-        </div>
-    </div>
-
-    <!-- Left and right controls/icons -->
-    <button class="carousel-control-prev" type="button" data-bs-target="#demo" data-bs-slide="prev">
-        <span class="carousel-control-prev-icon"></span>
-    </button>
-    <button class="carousel-control-next" type="button" data-bs-target="#demo" data-bs-slide="next">
-        <span class="carousel-control-next-icon"></span>
-    </button>
-</div>
-{% endblock %}
