--- QLHS/QLHocSinh/templates/setup-class.html
@@ -1,8 +1,0 @@
-{% extends "base.html" %}
-
-
-
-{% block content %}
-<h1 class="text-center heading">LẬP DANH SÁCH LỚP</h1>
-
-{% endblock %}
