--- QLHS/README.md
@@ -1,1 +1,0 @@
-# QuanLyHocSinh
