--- QLHS/requirements.txt
@@ -1,22 +1,0 @@
-certifi==2022.12.7
-click==8.1.3
-cloudinary==1.32.0
-colorama==0.4.6
-Flask==2.2.3
-Flask-Admin==1.6.1
-Flask-Login==0.6.2
-Flask-SQLAlchemy==3.0.3
-greenlet==2.0.2
-importlib-metadata==6.0.0
-itsdangerous==2.1.2
-Jinja2==3.1.2
-MarkupSafe==2.1.2
-PyMySQL==1.0.2
-six==1.16.0
-SQLAlchemy==2.0.4
-typing_extensions==4.5.0
-urllib3==1.26.15
-Werkzeug==2.2.3
-WTForms==3.0.1
-zipp==3.14.0
-
