+++ QLHocSinh/__init__.py
@@ -6,11 +6,12 @@ import cloudinary
 app = Flask(__name__)
 app.secret_key = '#%^&(*$%^&(78678675$%&^&$^%*&^%&*^'
 
-app.config["SQLALCHEMY_DATABASE_URI"] = "mysql+pymysql://root:Chuyende01htttql@localhost/qlhs?charset=utf8mb4"
+app.config["SQLALCHEMY_DATABASE_URI"] = "mysql+pymysql://root:123456@localhost/qlhs1?charset=utf8mb4"
 app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = True
 
 db = SQLAlchemy(app)
 login = LoginManager(app)
+
 
 cloudinary.config(cloud_name='dkcq9dgui', api_key='586758563261442', api_secret='VUEoTXbDRMiDSc9PXRcu_dqEoe0')
