+++ QLHocSinh/index.py
@@ -1,21 +1,17 @@
-from flask import render_template, request, redirect, session, Flask, url_for
+from flask import render_template, request, redirect, session, Flask
+from QLHocSinh import app, dao, admin
+from QLHocSinh import login
 from flask_login import login_user, login_required, logout_user
-from QLHocSinh.admin import *
-from QLHocSinh import app, dao, admin, login
 import cloudinary.uploader
-import utilities
+from QLHocSinh.admin import *
 
 @app.route("/")
 def index():
-    return render_template("index.html")
+    return render_template("dashboard.html")
 
 @login.user_loader
 def user_load(user_id):
     return dao.get_user_by_id(user_id)
-
-# @login.user_loader
-# def user_load(id):
-#     return utilities.get_user_by_id(id=id)
 
 @app.route('/login')
 def my_login():
@@ -46,29 +42,54 @@ def user_login():
 
     return render_template('login.html', error_msg=error_msg)
 
-# @app.route("/login", methods=['get', 'post'])
-# def my_login():
-#     err_msg = ''
-#     if request.method.__eq__('POST'):
-#         username = request.form.get('username')
-#         password = request.form.get('password')
-#
-#         user = utilities.check_login(username=username, password=password)
-#         if user:
-#             login_user(user=user)
-#             return redirect(url_for('index'))
-#         else:
-#             err_msg = 'Tên đăng nhập hoặc mật khẩu không chính xác'
-#     return render_template('login.html', err_msg=err_msg)
 
 @app.route("/logout")
 def my_logout():
     logout_user()
-    return redirect(url_for('my_login'))
+    return redirect('/login')
 
-@app.route("/add-students")
+# @app.route('/add-students', methods=['GET', 'POST'])
+@app.route('/add-students')
+@login_required
 def add_students():
+    if current_user.user_role == UserRole.NV:
+        if request.args.get('err_msg'):
+            return render_template("add-students.html", err_msg=request.args.get('err_msg'))
+
     return render_template("add-students.html")
+
+@app.route('/add-students', methods=['GET', 'POST'])
+@login_required
+def out_student():
+    if current_user.user_role == UserRole.QTV:
+        if request.method.__eq__('POST'):
+            first_name = request.form.get('ten')
+            last_name = request.form.get('hoten')
+            sex = request.form.get('gioitinh')
+            bday = request.form.get('ngaysinh')
+            address = request.form.get('diachi')
+            phone = request.form.get('sdt')
+            email = request.form.get('email')
+
+        dao.add_student(first_name=first_name,
+                                    last_name=last_name,
+                                    sex=sex,
+                                    bday=bday,
+                                    address=address,
+                                    phone=phone,
+                                    email=email)
+        info_student = {
+                'ten': first_name,
+                'hoten': last_name,
+                'gioitinh': 'Nam' if sex == 'male' else 'Nữ',
+                'ngaysinh': 'Ngày ' + bday.split('-')[2] + ' Tháng ' + bday.split('-')[1] + ' Năm ' + bday.split('-')[0],
+                'diachi': address,
+                'sdt': phone,
+                'email': email
+            }
+        return render_template("out-student.html", info_student=info_student)
+    else:
+        return redirect("/")
 
 @app.route("/setup-class")
 def setup_class():
@@ -78,8 +99,38 @@ def setup_class():
 def arrange_class():
     return render_template("arrange-class.html")
 
+# @app.route("/students-marks")
+# @login_required
+# def students_marks():
+#     course_id = request.args.get('course_id')
+#     keyword = request.args.get('keyword')
+#     classes = dao.get_classes_of_teacher(current_user.id)
+#     if keyword:
+#         filtered_classes = []
+#         for c in classes:
+#             for i in c:
+#                 if keyword.lower() in str(i).lower():
+#                     if c not in filtered_classes:
+#                         filtered_classes.append(c)
+#
+#         return render_template("students-marks.html", classes=filtered_classes)
+#
+#     if course_id:
+#         if dao.check_teacher_access(user_id=current_user.id, course_id=course_id):
+#             course = dao.get_course_info(course_id)
+#             dao.create_all_mark_records(course_id=course_id) # Tao bang diem khi vao nhap diem
+#             marks = utilities.get_mark_by_course_id(course_id=course_id)
+#
+#             return render_template("students-marks.html",
+#                                    marks=marks,
+#                                    course=course,
+#                                    classes=classes)
+#         else:
+#             return redirect("/")
+#     else:
+#         return render_template("students-marks.html", classes=classes)
+
+
 if __name__ == '__main__':
-    from QLHocSinh.admin import *
-    # app.run(debug=True)
-    app.run()
+    app.run(debug=True)
