+++ QLHocSinh/templates/base.html
@@ -38,39 +38,21 @@
                     <a class="nav-link" href="/arrange-class">Điều chỉnh lớp học</a>
                 </li>
                 <li class="nav-item">
+                    <a class="nav-link" href="/students-marks">Quản lý điểm</a>
+                </li>
+                <li class="nav-item">
                     <a class="nav-link" href="/logout">Đăng xuất</a>
                 </li>
                 {% else %}
                 <li class="nav-item">
-                    <a class="nav-link" href="/login">Đăng nhập</a>
+                        <a class="nav-link" href="/login">Đăng nhập</a>
                 </li>
                 {% endif %}
             </ul>
-            <form class="d-flex" action="/">
+             <form class="d-flex" action="/">
                 <input class="form-control me-2" name="kw" type="text" placeholder="Nhập tên...">
                 <button class="btn btn-primary" type="submit">Tìm</button>
-            </form>
-            <ul class="navbar-nav">
-                {% if current_user.is_authenticated %}
-                    <li class="nav-item">
-                        <a class="nav-link" href="{{ url_for('index') }}">
-                            {% if current_user.avatar %}
-                                <img src="{{ current_user.avatar }}" alt="{{ current_user.username }}" width="30">
-                            {% else %}
-                                <img src="{{ url_for('static', filename='images/Avatar.png') }}" alt="{{ current_user.username }}" width="30">
-                            {% endif %}
-                            {{ current_user.username }}
-                        </a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="{{ url_for('my_logout') }}">Đăng xuất</a>
-                    </li>
-                {% else %}
-                    <li class="nav-item">
-                        <a class="nav-link" href="{{ url_for('my_login') }}">Đăng nhập</a>
-                    </li>
-                {% endif %}
-            </ul>
+             </form>
         </div>
     </div>
 </nav>
