+++ QLHocSinh/templates/layout/base-common.html
@@ -25,53 +25,41 @@
                 <span class="navbar-toggler-icon"></span>
             </button>
             <div class="collapse navbar-collapse" id="collapsibleNavbar">
-                <ul class="navbar-nav me-auto">
-<!--                    <li class="nav-item">-->
-<!--                        <a class="nav-link" href="/">Trang chủ</a>-->
-<!--                    </li>-->
-<!--                    {% if current_user.is_authenticated %}-->
-<!--                    <li class="nav-item">-->
-<!--                        <a class="nav-link text-danger" href="/">-->
-<!--                            <img src="{{ current_user.avatar }}" width="40" class="rounded-circle" />{{ current_user.name }}-->
-<!--                        </a>-->
-<!--                    </li>-->
-<!--                    <li class="nav-item">-->
-<!--                        <a class="nav-link" href="/add-students">Tiếp nhận học sinh</a>-->
-<!--                    </li>-->
-<!--                    <li class="nav-item">-->
-<!--                        <a class="nav-link" href="/setup-class">Lập danh sách lớp</a>-->
-<!--                    </li>-->
-<!--                    <li class="nav-item">-->
-<!--                        <a class="nav-link" href="/arrange-class">Điều chỉnh lớp học</a>-->
-<!--                    </li>-->
-<!--                    <li class="nav-item">-->
-<!--                        <a class="nav-link" href="/logout">Đăng xuất</a>-->
-<!--                    </li>-->
-<!--                    {% else %}-->
-<!--                    <li class="nav-item">-->
-<!--                        <a class="nav-link" href="/login">Đăng nhập</a>-->
-<!--                    </li>-->
-<!--                    {% endif %}-->
-                </ul>
+                <ul class="navbar-nav me-auto"></ul>
                 <ul class="navbar-nav">
+                    <li class="nav-item">
+                        <a class="nav-link" href="/">Trang chủ</a>
+                    </li>
                     {% if current_user.is_authenticated %}
-                        <li class="nav-item">
-                            <a class="nav-link" href="{{ url_for('index') }}">
-                                {% if current_user.avatar %}
-                                    <img src="{{ current_user.avatar }}" alt="{{ current_user.username }}" width="30">
-                                {% else %}
-                                    <img src="{{ url_for('static', filename='images/Avatar.png') }}" alt="{{ current_user.username }}" width="30">
-                                {% endif %}
-                                {{ current_user.username }}
-                            </a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link" href="{{ url_for('my_logout') }}">Đăng xuất</a>
-                        </li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="/add-students">Tiếp nhận học sinh</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="/setup-class">Lập danh sách lớp</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="/arrange-class">Điều chỉnh lớp học</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="/students-marks">Quản lý điểm</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="{{ url_for('index') }}">
+                            {% if current_user.avatar %}
+                                <img src="{{ current_user.avatar }}" alt="{{ current_user.username }}" width="26" class="rounded-circle">
+                            {% else %}
+                                <img src="{{ url_for('static', filename='images/Avatar.png') }}" alt="{{ current_user.username }}" width="26" class="rounded-circle">
+                            {% endif %}
+                            {{ current_user.username }}
+                        </a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="{{ url_for('my_logout') }}">Đăng xuất</a>
+                    </li>
                     {% else %}
-                        <li class="nav-item">
-                            <a class="nav-link" href="{{ url_for('my_login') }}">Đăng nhập</a>
-                        </li>
+                    <li class="nav-item">
+                            <a class="nav-link" href="/login">Đăng nhập</a>
+                    </li>
                     {% endif %}
                 </ul>
             </div>
