+++ QLHocSinh/templates/login.html
@@ -11,7 +11,7 @@
     </div>
     {% endif %}
     <!--  them attribute enctype="multipart/form-data" neu them image-->
-    <form method="post" action="{{ url_for('my_login') }}">
+    <form method="post" action="{{ url_for('add-students') }}">
         <div class="form-group mb-3 mt-3">
             <label for="username" class="form-label">Tên đăng nhập:</label>
             <input type="text" class="form-control" id="username" placeholder="Tên đăng nhập" name="username">
