+++ QLHocSinh/templates/setup-class.html
@@ -1,5 +1,7 @@
 {% extends "base.html" %}
 
+
+
 {% block content %}
 <h1 class="text-center heading">LẬP DANH SÁCH LỚP</h1>
